server/pyproject.toml
[project]
name = "learnx-server"
version = "0.1.0"
description = "Fastapi backend for LearnX"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.3",
]
